requirements.txt
pycrypto
bcrypt==3.1.6
cffi==1.12.2
Click==7.0
dnspython==1.16.0
eventlet==0.24.1
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
monotonic==1.5
pycparser==2.19
python-engineio==3.5.1
python-socketio==4.0.0
six==1.12.0
SQLAlchemy==1.3.2
Werkzeug==0.15.2
